Deletion of private products and uploaded files

+++ tao-data-services/src/main/java/ro/cs/tao/services/entity/impl/FileStorageService.java
@@ -23,9 +23,14 @@ import org.springframework.web.multipart.MultipartFile;
 import ro.cs.tao.component.SystemVariable;
 import ro.cs.tao.configuration.ConfigurationManager;
 import ro.cs.tao.eodata.AuxiliaryData;
+import ro.cs.tao.eodata.EOProduct;
+import ro.cs.tao.eodata.VectorData;
+import ro.cs.tao.eodata.enums.Visibility;
 import ro.cs.tao.persistence.PersistenceManager;
+import ro.cs.tao.persistence.exception.PersistenceException;
 import ro.cs.tao.security.SessionStore;
 import ro.cs.tao.services.interfaces.StorageService;
+import ro.cs.tao.utils.FileUtils;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -34,6 +39,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.stream.Stream;
 
 @Service("storageService")
@@ -72,13 +78,57 @@ public class FileStorageService implements StorageService<MultipartFile> {
 
     @Override
     public void remove(String fileName) throws IOException{
+        if (fileName == null || fileName.isEmpty()) {
+            throw new IOException("Invalid argument (empty)");
+        }
         if (fileName.contains("..")) {
             // This is a security check
             throw new IOException( "Cannot remove file with relative path outside user directory " + fileName);
         }
-        Path filePath = SessionStore.currentContext().getUploadPath().resolve(fileName);
-        Files.delete(filePath);
-        persistenceManager.removeAuxiliaryData(SessionStore.currentContext().getWorkspace().relativize(filePath).toString());
+        Path uploadPath = SessionStore.currentContext().getUploadPath();
+        if (fileName.equals(uploadPath.toString()) ||
+                fileName.startsWith(SystemVariable.SHARED_WORKSPACE.value()) ||
+                fileName.startsWith(SystemVariable.SHARED_FILES.value())) {
+            throw new IOException( "Operation not allowed");
+        }
+        if (fileName.startsWith("files")) {
+            Path filePath = uploadPath.resolve(fileName);
+            Files.delete(filePath);
+            persistenceManager.removeAuxiliaryData(SessionStore.currentContext().getWorkspace().relativize(filePath).toString());
+        } else { // it has to be a product folder
+            Path filePath = SessionStore.currentContext().getWorkspace().resolve(fileName);
+            List<EOProduct> products = persistenceManager.getEOProducts(filePath.toUri().toString());
+            if (products == null || products.size() == 0) {
+                List<VectorData> vectorProducts = persistenceManager.getVectorDataProducts(filePath.toUri().toString());
+                if (vectorProducts.size() > 1) {
+                    throw new IOException("Cannot remove vector products");
+                }
+                VectorData product = vectorProducts.get(0);
+                if (Visibility.PUBLIC.equals(product.getVisibility())) {
+                    throw new IOException("Cannot remove a public product. Please reduce its visibility first.");
+                }
+                FileUtils.deleteTree(filePath.toFile());
+                try {
+                    persistenceManager.remove(product);
+                } catch (PersistenceException e) {
+                    throw new IOException(e);
+                }
+            } else {
+                if (products.size() > 1) {
+                    throw new IOException("Cannot remove raster products");
+                }
+                EOProduct product = products.get(0);
+                if (Visibility.PUBLIC.equals(product.getVisibility())) {
+                    throw new IOException("Cannot remove a public product. Please reduce its visibility first.");
+                }
+                FileUtils.deleteTree(filePath.toFile());
+                try {
+                    persistenceManager.remove(product);
+                } catch (PersistenceException e) {
+                    throw new IOException(e);
+                }
+            }
+        }
     }
 
     @Override
